chore: added missing ArangoDB DatabaseManager producer
