--- arangodb/deployment/src/main/java/io/quarkiverse/jnosql/document/arangodb/deployment/Processor.java
@@ -1,24 +1,0 @@
-package io.quarkiverse.jnosql.document.arangodb.deployment;
-
-import org.eclipse.jnosql.databases.arangodb.communication.QuarkusArangoDBDocumentConfiguration;
-
-import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
-import io.quarkus.deployment.annotations.BuildProducer;
-import io.quarkus.deployment.annotations.BuildStep;
-import io.quarkus.deployment.builditem.FeatureBuildItem;
-
-class Processor {
-
-    private static final String FEATURE = "jnosql-arangodb";
-
-    @BuildStep
-    FeatureBuildItem feature() {
-        return new FeatureBuildItem(FEATURE);
-    }
-
-    @BuildStep
-    void build(BuildProducer<AdditionalBeanBuildItem> additionalBeanProducer) {
-        additionalBeanProducer.produce(AdditionalBeanBuildItem.unremovableOf(QuarkusArangoDBDocumentConfiguration.class));
-    }
-
-}
